+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Utils.hs
@@ -10,38 +10,11 @@ module Test.Cardano.Ledger.Core.Utils (
 )
 where
 
-import Cardano.Ledger.BaseTypes (
-  EpochSize (..),
-  Globals (..),
-  Network (..),
-  knownNonZeroBounded,
-  mkActiveSlotCoeff,
- )
 import Cardano.Ledger.Core
 import Cardano.Ledger.Hashes (unsafeMakeSafeHash)
 import Cardano.Ledger.TxIn (TxIn, mkTxInPartial)
-import Cardano.Slotting.EpochInfo (fixedEpochInfo)
-import Cardano.Slotting.Time (SystemStart (..), mkSlotLength)
-import Data.Time.Clock.POSIX (posixSecondsToUTCTime)
 import Test.Cardano.Ledger.Binary.Random (mkDummyHash)
 import Test.Cardano.Ledger.Common
-import Test.Cardano.Ledger.Core.Rational (unsafeBoundRational)
-
-testGlobals :: Globals
-testGlobals =
-  Globals
-    { epochInfo = fixedEpochInfo (EpochSize 100) (mkSlotLength 1)
-    , slotsPerKESPeriod = 20
-    , stabilityWindow = 33
-    , randomnessStabilisationWindow = 33
-    , securityParameter = knownNonZeroBounded @10
-    , maxKESEvo = 10
-    , quorum = 5
-    , maxLovelaceSupply = 45 * 1000 * 1000 * 1000 * 1000 * 1000
-    , activeSlotCoeff = mkActiveSlotCoeff . unsafeBoundRational $ 0.9
-    , networkId = Testnet
-    , systemStart = SystemStart $ posixSecondsToUTCTime 0
-    }
 
 mkDummySafeHash :: forall a. Int -> SafeHash a
 mkDummySafeHash = unsafeMakeSafeHash . mkDummyHash @HASH